# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LeptonPad API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/gateway/{contentId}:
    get:
      operationId: payGatewayUnlock
      tags: [payments]
      summary: Unlock content via x402 Gateway nanopayment
      parameters:
        - name: contentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unlock result after verified payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnlockResult"
        "402":
          description: Payment required — x402 PAYMENT-REQUIRED header included
        "401":
          description: Unauthorized
  /payments/unlock:
    post:
      operationId: unlockContent
      tags: [payments]
      summary: Unlock content with USDC payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnlockInput"
      responses:
        "200":
          description: Unlock result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnlockResult"
        "400":
          description: Payment failed
        "401":
          description: Unauthorized
  /payments/check/{contentId}:
    get:
      operationId: checkContentAccess
      tags: [payments]
      summary: Check if current user has paid for content
      parameters:
        - name: contentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Access status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessStatus"
...
components:
  schemas:
...
    UnlockInput:
      type: object
      properties:
        contentId:
          type: integer
        txHash:
          type: ["string", "null"]
      required:
        - contentId
    UnlockResult:
      type: object
      properties:
        success:
          type: boolean
        paymentId:
          type: integer
        contentId:
          type: integer
        amountPaid:
          type: number
        creatorReceives:
          type: number
        settledAt:
          type: string
      required:
        - success
        - paymentId
        - contentId
        - amountPaid
        - creatorReceives
        - settledAt
    AccessStatus:
      type: object
      properties:
        hasAccess:
          type: boolean
        paymentId:
          type: ["integer", "null"]
      required:
        - hasAccess
        - paymentId
...